Attempt to improve missing data error handling

+++ code/tedana-rf1-hpc.sh
@@ -11,6 +11,7 @@ cd $PBS_O_WORKDIR
 maindir=/home/tun31934/work/rf1-sra-data
 scriptdir=$maindir/code
 logdir=$maindir/logs
+prepdir=$maindir/derivatives/fmriprep
 mkdir -p $logdir
 
 rm -f $logdir/cmd_tedana_${PBS_JOBID}.txt
@@ -19,33 +20,37 @@ touch $logdir/cmd_tedana_${PBS_JOBID}.txt
 for sub in ${subjects[@]}; do
 	for task in "sharedreward" "trust" "ugr" "doors" "socialdoors"; do
 		for run in 1 2; do
-		
-			# prepare inputs and outputs; don't run if data is missing, but log missingness
+
+			# prepare inputs and outputs
 			prepdir=${maindir}/derivatives/fmriprep/sub-${sub}/func
 			echo1=${prepdir}/sub-${sub}_task-${task}_run-${run}_echo-1_part-mag_desc-preproc_bold.nii.gz
 			echo2=${prepdir}/sub-${sub}_task-${task}_run-${run}_echo-2_part-mag_desc-preproc_bold.nii.gz
 			echo3=${prepdir}/sub-${sub}_task-${task}_run-${run}_echo-3_part-mag_desc-preproc_bold.nii.gz
 			echo4=${prepdir}/sub-${sub}_task-${task}_run-${run}_echo-4_part-mag_desc-preproc_bold.nii.gz
 			outdir=${maindir}/derivatives/tedana/sub-${sub}
-			if [ ! -e $echo1 ]; then
-				echo "missing ${echo1}"
-				echo "missing ${echo1}" >> $scriptdir/missing-tedanaInput.log
-				exit
-			fi
-			mkdir -p $outdir
 			
+			# Check for the presence of all echo files
+			if [ ! -e $echo1 ] || [ ! -e $echo2 ] || [ ! -e $echo3 ] || [ ! -e $echo4 ]; then
+				echo "Missing one or more files for sub-${sub}, task-${task}, run-${run}" >> $scriptdir/missing-tedanaInput.log
+				echo "Skipping sub-${sub}, task-${task}, run-${run}" >> $logdir/cmd_tedana_${PBS_JOBID}.txt
+				continue
+			fi
+
+			mkdir -p $outdir
+
 			# run tedana
-			tedana -d $echo1 $echo2 $echo3 $echo4 \
+			echo "Running tedana -d $echo1 $echo2 $echo3 $echo4 \
 			-e 0.0138 0.03154 0.04928 0.06702 \
 			--out-dir $outdir \
 			--prefix sub-${sub}_task-${task}_run-${run} \
 			--convention bids \
 			--fittype curvefit \
-			--overwrite >> $logdir/cmd_fmriprep_${PBS_JOBID}.txt
-			
+			--overwrite" >> $logdir/cmd_tedana_${PBS_JOBID}.txt
+
+
 			# clean up and save space
 			rm -rf ${outdir}/sub-${sub}_task-${task}_run-${run}_*.nii.gz
-			
+
 		done
 	done
 done
